requirements.txt
pillow
pandas
numpy
matplotlib
seaborn
pymysql